+++ src/gui/web-components/util/NotificationWebComponent.ts
@@ -7,29 +7,18 @@ import {
 import { AbstractStateSubscriberWebComponent } from "./AbstractStateSubscriberWebComponent.js";
 
 export class NotificationWebComponent extends AbstractStateSubscriberWebComponent {
-    private hideTimer: ReturnType<typeof setTimeout> | null = null;
     private readonly hideDelay = 3_000;
-
-    private elements: {
-        output: HTMLOutputElement;
-    };
+    private readonly animiationDuration = 500;
 
     constructor() {
         super();
-        this.elements = {
-            output: document.createElement("output"),
-        };
-        this.elements.output.role = "alert";
-        this.elements.output.ariaLive = "assertive";
-        this.elements.output.ariaAtomic = "true";
-        this.appendChild(this.elements.output);
+        this.role = "alert";
+        this.ariaLive = "assertive";
+        this.ariaAtomic = "true";
     }
 
     connectedCallback() {
         document.body.addEventListener(NOTIFICATION_EVENT_IDENTIFIER, this.handleNotification);
-        this.addEventListener("click", () => {
-            this.elements.output.classList.toggle("is-visible");
-        });
 
         this.subscribeToStateUpdates((state) => {
             if (state.isMissingMainMcf) {
@@ -56,18 +45,23 @@ export class NotificationWebComponent extends AbstractStateSubscriberWebComponen
         this.handleNotificationDetails(details);
     };
 
-    handleNotificationDetails<T>(details: NotificationEventPayload<T>) {
+    private handleNotificationDetails<T>(details: NotificationEventPayload<T>) {
         this.log(details);
 
-        this.elements.output.classList.remove("info", "success", "error", "waiting");
-        this.elements.output.classList.add(details.type, "is-visible");
-        this.elements.output.innerText = this.getEmoji(details.type) + details.message;
+        const output = document.createElement("output");
+        output.innerText = this.getEmoji(details.type) + details.message;
+        this.appendChild(output);
 
-        if (this.hideTimer !== null) {
-            clearTimeout(this.hideTimer);
-        }
-        this.hideTimer = setTimeout(() => {
-            this.elements.output.classList.remove(details.type, "is-visible");
+        setTimeout(() => {
+            output.classList.add(details.type, "is-visible");
+        }, 1);
+
+        setTimeout(() => {
+            output.classList.remove("is-visible");
+
+            setTimeout(() => {
+                output.remove();
+            }, this.animiationDuration);
         }, this.hideDelay);
     }