+++ types/gui/web-components/util/NotificationWebComponent.d.ts
@@ -1,14 +1,12 @@
-import { type NotificationEventPayload } from "../../renderer/notificationEventHandler.js";
 import { AbstractStateSubscriberWebComponent } from "./AbstractStateSubscriberWebComponent.js";
 export declare class NotificationWebComponent extends AbstractStateSubscriberWebComponent {
-    private hideTimer;
     private readonly hideDelay;
-    private elements;
+    private readonly animiationDuration;
     constructor();
     connectedCallback(): void;
     disconnectedCallback(): void;
     private handleNotification;
-    handleNotificationDetails<T>(details: NotificationEventPayload<T>): void;
+    private handleNotificationDetails;
     private getEmoji;
     private log;
     static registerElement(): void;